implement login, session management and protected routes (untested)

+++ recetly-back/controllers/recipe.js
@@ -2,6 +2,8 @@ import { RecipeModel } from "../models/recipe.js"
 
 export class RecipeController {
     static getAll = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
         try{
         const recipes = await RecipeModel.getAll()
         res.json(recipes)
@@ -14,9 +16,11 @@ export class RecipeController {
     
     
     static getById = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
         try{
-        const { id }  = req.params
-        const recipe = await RecipeModel.getById(id)
+        const { user_id }  = user
+        const recipe = await RecipeModel.getById(user_id)
         res.json(recipe)
         } catch (error) {
             console.error(error);
@@ -24,9 +28,11 @@ export class RecipeController {
         }
     }
     static postRecipe = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
         try {
             const input = req.body;
-            await RecipeModel.postRecipe(input); 
+            await RecipeModel.postRecipe(input, user.user_id); 
             res.status(201).end(); 
         } catch (error) {
             console.error(error);
@@ -34,6 +40,8 @@ export class RecipeController {
         }
     }
     static deleteRecipe = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
         try{
         const { id } = req.params
         await RecipeModel.deleteById(id)
@@ -43,7 +51,20 @@ export class RecipeController {
             res.status(400).json({ error: error.message });
         }
     }
+    static deleteRecipeUser = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
+        try{
+        await RecipeModel.deleteByUser(user.user_id)
+        res.status(200).end()
+        } catch (error) {
+            console.error(error);
+            res.status(400).json({ error: error.message });
+        }
+    }
     static updateRecipe = async (req, res) => {
+        const { user } = req.session
+        if(!user) return res.status(403).send('Access not authorized')
         try{
         const { id } = req.params
         const input = req.body
